+++ CHANGELOG.md
@@ -6,8 +6,8 @@
 
 ### ✨ Features
 
-* enableOwnerStacks feature to disable owner stacks
+- enableOwnerStacks feature to disable owner stacks
 
 ### 📚 Documentation
 
-* README ([f381743](https://github.com/mdjastrzebski/react-test-config/commit/f38174365b1203710bfb08007cd22f9c46bd8e53))
+- README ([f381743](https://github.com/mdjastrzebski/react-test-config/commit/f38174365b1203710bfb08007cd22f9c46bd8e53))